refactor: break things out a bit
