+++ typescript/optics-provider/src/contracts.ts
@@ -11,13 +11,9 @@ import {
 type Address = string;
 type ProviderOrSigner = ethers.providers.Provider | ethers.Signer;
 
-interface ReplicaInfo {
-  domain: number;
-  address: Address;
-}
-
 export abstract class Contracts {
   readonly original: Object;
+
   constructor(data: Object) {
     this.original = data;
   }
@@ -42,67 +38,5 @@ export abstract class Contracts {
   ): T {
     let file = fs.readFileSync(filepath);
     return new this(file);
-  }
-}
-
-export class BridgeContracts {
-  bridgeRouter: BridgeRouter;
-  ethHelper: ETHHelper;
-
-  constructor(br: Address, ethHelper: Address, signer?: ethers.Signer) {
-    this.bridgeRouter = new BridgeRouter__factory(signer).attach(br);
-    this.ethHelper = new ETHHelper__factory(signer).attach(ethHelper);
-  }
-
-  connect(signer: ethers.Signer) {
-    this.bridgeRouter = this.bridgeRouter.connect(signer);
-    this.ethHelper = this.ethHelper.connect(signer);
-  }
-
-  fromObject(data: any, signer?: ethers.Signer) {
-    if (!data.bridgeRouter || !data.ethHelper) {
-      throw new Error("missing address");
-    }
-
-    const br = data.bridgeRouter.proxy ?? data.bridgeRouter;
-    const eh = data.bridgeRouter.proxy ?? data.bridgeRouter;
-
-    return new BridgeContracts(br, eh);
-  }
-
-  toObject(): Object {
-    return {
-      bridgeRouter: this.bridgeRouter.address,
-      ethHelper: this.ethHelper.address,
-    };
-  }
-}
-
-export class CoreContracts {
-  home: core.Home;
-  replicas: Record<number, core.Replica>;
-
-  constructor(home: Address, replicas: ReplicaInfo[], signer?: ethers.Signer) {
-    this.home = new core.Home__factory(signer).attach(home);
-
-    this.replicas = [];
-    replicas.forEach((replica) => {
-      this.replicas[replica.domain] = new core.Replica__factory(signer).attach(
-        replica.address
-      );
-    });
-  }
-
-  toObject(): Object {
-
-    const replicas: Record<number, string> = {};
-    Object.entries(this.replicas).forEach(([k, v]) => {
-      replicas[k] = v.address;
-    });
-
-    return {
-      home: this.home.address,
-      replicas:
-    }
   }
 }
